requirements.txt
Flask==3.1.0
opencv-python==4.6.0.66
requests==2.31.0
python-dotenv==1.0.1